.idea/requirements.txt
PySide6==6.7.2
qasync==0.24.0
websockets==12.0
PyYAML==6.0.2
pycryptodome==3.20.0
